Fix: support nested structures for `FieldOffsetGetters` (#95)
Problem: the derive macro for `FieldOffsetGetters` assumes a size of 1
for each field of a struct, even if they are structures that span
multiple felts.

Solution: take this into account in the offset computation.
Deriving `FieldOffsetGetters` now adds a `cairo_size() -> usize` method
that computes the size of the struct in terms of felts. This method is
called for each struct encountered in the derive macro.