Internal: introduce the `Hash` type for hashes
Problem: we did not use a custom type when porting most of the code,
resulting in hashes being `Vec<u8>` all over the place. This is
inconvenient and error-prone.

Solution: introduce the `Hash` type to encapsulate all hashes. With
this, hashes are now type-safe and provide implementations of `From` for
common types in Cairo VM and Starknet API.