Feature: library_call syscall for Cairo 0 and 1
Added support and tests for the library_call syscall in deprecated and
regular syscalls.

Refactored the code to use the same function for calling a contract
directly and using `library_call`, similarly to what is done in the
Python VM.

Added implementations for all the deprecated "call" syscalls as well:
* `delegate_call`
* `delegate_l1_handler`
* `library_call_l1_handler`

as their implementation relies on the same logic. No tests could be
added for these at the current time.