Fix: load contract and class changes in OS output (#235)
Problem: the `contracts` and `classes` fields of the OS output are not
deserialized correctly. We need to deserialize a full structure and not
just felts.

Solution: add specific deserializers for contract and class changes.
Additionally, add the OS output to the return value when running the OS
in order to check specific output values in changes.