Feature: compile Sierra classes in tests
Problem: the Starknet API only provides Sierra classes and no CASM. We
need both to run the OS. So far we used pre-compiled contract classes
but it would be better to use the same flow for production and for
tests.

Solution: compile Sierra classes in the Cairo 1 contract loader for
tests. Removed all pre-compiled CASM files, except for one that is used
to test serialization/deserialization of compiled class facts.