Internal: make the project a Cargo workspace (#271)
Problem: the project is getting more complex, with the need for
additional crates for compatibility with RPC providers etc. We want to
split dependencies for specific use cases and isolate needs.

Solution: turn the project into a Cargo workspace. Every library
and binary is now located in the `crates` directory. Dependencies are
now managed in the top-level `Cargo.toml` file.