Fix: include compiled OS in the library binary (#276)
Problem: the OS was configured as a relative path inside the
`starknet-os` library, leading to potential issues when using
the library in different contexts.

Solution: load the compiled OS as a constant. Note that this makes the
library quite heavy as the OS takes around 32MB. We believe this is
acceptable for a first iteration.