Fix: copy-paste issues in the implementation of `get_block_timestamp()`
Problem: the implementation of the `get_block_timestamp()` syscall was
hastily copied from the implementation of `get_block_number()` and uses
the wrong types. There is no functional problem because the types used
have the same sizes, but this needs to be cleaned up.

Solution: define the request/response types and use them.