Feature: generic CASM class (#290)
Problem: we use different CASM contract classes depending on the use
case. We serialize/deserialize with the `cairo-lang` implementation, we
reexecute txs with the Blockifier implementation and we compute class
hashes through Pathfinder. Currently, we convert explicitly between
these types/crates. This makes the code harder to read and maintain.

Solution: introduce a generic class that is able to convert to/from all
the types we use under the hood. This puts a lid on the problem and
reduces the amount of code to change if things improve in the ecosystem.