Pathfinder integration with manual block reexecution (#287)
* Internal: isolate Blockifier compatibility from `SharedState`

Problem: the `SharedState` struct currently has a `contract_addresses`
hashset that maps all the contracts in the state. We do not want to
require an exhaustive set of all contracts when using real block data
from Pathfinder.

Solution: isolate the Blockifier-related code of `SharedState` (state
diff application + type conversions) to a new `TestSharedState` struct
dedicated to the task. `SharedState` is now a base struct that aims to
be compatible with all use cases (transaction reexecution for tests,
fetch the trees from an RPC for production).

* wip

* wip

* rebase fix

* Make replay work again

* update storage proof, get state and storage proofs

* Start replacing starknet-replay with manual blockifier execution

* Implement StateReader with Pathfinder RPC

* Attempt to guess more facts to write from RPC

* Fall back correctly to RPC when DictStateReader is missing entry

* Remove starknet-replay

* Fix most compile warnings

* Create state commitment info

* Query API with class_hash, not contract_address

* Remove DictStateWithRpc

* Revert "Internal: isolate Blockifier compatibility from `SharedState`"

This reverts commit 0a8f8d587c242be2d3e9ccac85da80dd4d91ac32.

* Construct SharedState properly

* fmt

* Fix some clippy lints

* More clippy fixes

* Remove unused dependency

* Reinstate deprecated class processing

---------

Co-authored-by: Olivier Desenfans <desenfans.olivier@gmail.com>