Feature: make OS per-contract storage customizable
Problem: we are facing issues integrating the model of
`OsSingleStarknetStorage` with Pathfinder, as we do not have a method to
traverse the tree to reach arbitrary leaves.

Solution: Make the OS generic on the per-contract storage type instead
of just the storage. This allows users to further customize the
behaviour of the OS when it comes to storage.