Fix: decompress starknet-core legacy class (#313)
Problem: the conversion from starknet-core `CompressedLegacyContractClass`
in `GenericDeprecatedContractClass` is untested and broken. This class
cannot be used interchangeably with other deprecated contract classes
because the `program` field is deserialized as a gzip bytes vector.

Solution: add conversion functions to decompress the program field and
generate a `LegacyContractClass` from it before storing it in the
generic class.