Feature: support Sierra 1.6 (#328)
* Feature: support Sierra 1.6

Problem: some contracts on Sepolia are built with Sierra 1.6, which is
not supported by the version of the CASM compiler we use. But bumping
the version of the compiler requires bumping cairo-vm and Blockifier as
well.

Solution: bump everything! We now use:
* cairo-vm 1.0.1
* cairo-lang 2.7.1
* blockifier 0.8.0-rc.2

With this, we can now compile any contract on Sepolia.

* Internal: remove deprecated `to_felt252` function

This function is no longer required as felt types now all rely on
starknet-types-core.

* Internal: remove deprecated `felt_vm2api` function

This function is no longer required as felt types now all rely on
starknet-types-core.

* Internal: remove deprecated `felt_api2vm` function

This function is no longer required as felt types now all rely on
starknet-types-core.

* clippy

* fmt

* Feature: support blocks requiring Starknet v0.13.2

* Internal: remove deprecated `felt_to_vm` function

This function is no longer required as felt types now all rely on
starknet-types-core.

* CI: add tests for DiffAssertValues resolution

* CI: regression test for diff assert values in contract subcall

* doc chain_id

* fix review: expect(todo)

* fix of the fix