Feature: upgrade OS to 0.13.2 (#337)
This commit is a combination of the following patches:

* Fix: make StarknetGeneralConfig compatible with 0.13.2

Problem: `StarknetGeneralConfig` is unable to load the default config
file for 0.13.2.

Solution: remove the now deprecated `constant_gas_price` field and set
`use_kzg_da` to true by default.

* Fix: use `all_cairo` layout in tests

Problem: the OS now requires the new builtins (`range_check96`,
`add_mod` and `mul_mod`) to be in the layout.

Solution: the only layout that provides all these builtins is
`all_cairo`. We now use this layout in all the integration tests.

* Internal: update unimplemented.rs

* Feature: implement SET_SHA256_SEGMENT_IN_SYSCALL_HANDLER hint

* Fix: OS_INPUT_TRANSACTIONS memory offset

Problem: the memory offset for this hint changed from 8 to 12.

* Feature: implement LOG_REMAINING_TXS hint

* Feature: FIND_ELEMENT hint

* Fix: update ENTER_CALL hint

The signature of enter_call was updated to rename the parameter. This
does not alter the behavior of the hint.

* Feature: FILL_HOLES_IN_RC96_SEGMENT hint

* Fix: update WRITE_USE_ZKG_DA_TO_MEM hint

The memory offset and the condition used to determine whether to use KZG
DA were modified in 0.13.2.

* Feature: WRITE_FULL_OUTPUT_TO_MEM hint

* Feature: minimal implementation of the CONFIGURE_KZG_MANAGER hint

We leave the `kzg_manager` part for later as this will probably conflict
with the KZG PR that is already open.

* Feature: SET_AP_TO_PREV_BLOCK_HASH and SET_AP_TO_NEW_BLOCK_HASH hints

* Fix: update SET_TREE_STRUCTURE hint

This hint is now executed only if
__serialize_data_availability_create_pages__ is true.

* Fix: update output deserialization

Many new fields were added, breaking the deserialization.

* Fix: make all tests compile

* Fix: use correct layout in prove_block

* Fix: clean up unimplemented.rs

* Fix: use cairo-lang 0.13.2 in CI