Fix: support `replace_class()` calls in invoke txs
Problem: some blocks fail on divergences in the global state tree caused
by calls to the `replace_class()` syscall.

Solution: use the class hash in block N-1 when building the
`ContractClass` objects instead of the current block, and fetch the
compiled class for both blocks to build the trees properly.