Fix: support writes to zero
Problem: we encounter an "Edge bottom not found in preimage" error when
running some blocks. This problem was diagnosed as writes to zero where
the last node of the storage proof was an edge node. A write to zero to
a specific address will clear the path to that leaf node, so the storage
proof for this key will always be incomplete. It turns out that the OS
will still look for the bottom node in the preimage to verify the node
hash.

Solution: if we encounter such a case, we now fetch an additional key
that is designed to follow the edge path. This way, we get the bottom
node in the preimage and the OS can compute the hash it wants to
compute.