Fix: support invoke txs that call the deploy syscall
Problem: the setup for `prove_block` assumes that we are always able to
fetch the class hash, class and storage proof for the contract addresses
found in invoke txs in the previous block. This is not always the case
as invoke txs are able to deploy contracts.

Solution: accept "contract not found" errors and storage proofs where
the `contract_data` field is missing. Note that we should implement a
way to double-check our expectations as this fix will accept an
incomplete storage proof in the current block as well, which is not
ideal.