Feature: rpc-client crate
Problems:
* We manage two types of clients, one for Starknet RPC endpoints and
  another for Pathfinder proof endpoints.
* We end up spawning way more clients than necessary because the
  starknet-rs client type does not implement Clone.

Solution: encapsulate all RPC-related objects in a new struct,
`RpcClient`. This struct lives in the new `rpc-client` crate.

* Moved all the code related to proofs in the `rpc-client` crate.
* Updated `rpc-replay` and `prove-block` to use the new RPC client.
* Added a layer to encapsulate the raw `reqwest` client used for
  Pathfinder proof endpoints in the new `PathfinderRpcClient` struct.

At this stage `RpcClient` still provides direct access to the Starknet
and Pathfinder RPC clients, and does not implement anything fancy (read:
caching). The goal of this PR is to clean up the code and encapsulate
functionalities.