Fix: return 0 as default storage value instead of error
Problem: similarly to what is happening for class hashes and nonces,
we sometimes fetch storage values for non-existing contracts (ex: when
the contract was deployed in the current block). When this happens, we
fail on a `ContractNotFound` error in the RPC state reader and
per-contract storage.

Solution: return 0 as default value when we do not find the storage.