fix: handle delegate_call syscalls correctly
Currently the delegate_syscall had been silenced with a noop storage.write.
This PR fixes that and adds a test in the run_os tests to ensure that the call is handled correctly

This PR also adds code_addresses to the map of accessed_storage_keys so that the os can successfully access the storage of contracts that were only invoked in a delegate_call.